Move jsontype to rowbinary mod
