+++ src/exporters/clickhouse/rowbinary/mod.rs
@@ -1,3 +1,5 @@
 pub(crate) use ser::serialize_into;
 
+pub mod json;
+
 mod ser;
